# CUDA: optimize FA for GQA + large batches